+++ platforms/kia_soul/firmware/steering/src/communications.cpp
@@ -130,8 +130,7 @@ static void process_chassis_state_1_report(
         const oscc_report_chassis_state_1_data_s * const chassis_state_1_data =
                 (oscc_report_chassis_state_1_data_s *) data;
 
-        float raw_angle =
-            (float) chassis_state_1_data->steering_wheel_angle * RAW_ANGLE_SCALAR;
+        int16_t raw_angle = chassis_state_1_data->steering_wheel_angle * RAW_ANGLE_SCALAR;
 
         g_steering_control_state.steering_angle =
             raw_angle * WHEEL_ANGLE_TO_STEERING_WHEEL_ANGLE_SCALAR;
